+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Remove the translation toolbar button
 - Move the proofreading button to the preview/check-in group
 - Add proofreading props
+- Add fullscreen toggle button
 
 ## 1.17.2 - 2024-02-11